Realization of web site for presentation of furniture sales
It is necessary to carry out a static presentation of the furniture shop. The site needs to be realized by creating a CMS (Content Management System) content update system. It is necessary to allow the site administrator full control over all textual content of the site, as well as control of uploading images and documents.